requirements.txt
slack_sdk
requests
beautifulsoup4
discord.py
pyyaml
selenium
pandas
typer